refactor: eliminate brightray::MainDelegate (#15333)
